open-unix: Implement with start-process
The old implementation with start-process-shell-command did not work on
Debian GNU/Linux 10 (buster) when xdg-open is used as command. Like
http://ergoemacs.org/emacs/emacs_dired_open_file_in_ext_apps.html, use
start-process with process-connection-type
nil. set-process-query-on-exit-flag nil to prevent Emacs from querying
whether the process should be killed when it exits.

Tested on Debian GNU/Linux 10 (buster) with evince and xdg-open.